refactor: move else folder to config folder
