+++ src/config/lexer.ts
